requirements.txt
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
transformers==4.46.3
sentence-transformers==3.3.1
huggingface-hub==0.26.2
numpy==2.1.3
scipy==1.14.1
scikit-learn==1.5.2
joblib==1.4.2
pillow==10.2.0
requests==2.32.3
tqdm==4.67.1
packaging==24.2
psutil==6.1.0
pyyaml==6.0.2
regex==2024.11.6
setuptools==70.0.0
Flask
faiss-cpu==1.9.0.post1
Flask==3.1.0